@@ -65,11 +65,11 @@ internal class AiCommenter
 			options.DeploymentId = parts[1];
 		}
 
-		var openAiService = new OpenAIService(options);
+		OpenAIService openAiService = new(options);
 
 
 		Models.Model model = arg.ParseResult.GetValueForOption(AiAssistorCommands.OpenAiModelOption);
-		List<ChatMessage> chatMessages = new();
+		List<ChatMessage> chatMessages = [];
 
 		BuildInitialInstruction(chatMessages, initialPrompt);
 
@@ -104,22 +104,22 @@ internal class AiCommenter
 		int tokenCount = 0;
 		double costCent = 0;
 
-		StringBuilder response = new();
 
+		List<AiCodeReviewComment> codeReviewComments = [];
 		Console.WriteLine("Requesting AI comments.");
 		bool allTokensSpend = false;
 		bool hasContent = false;
 		foreach (ChangePrompt input in validChangeInputs)
 		{
 			bool addToExisting =
-				input.Content.Length / 4 + chatMessages.Sum(c=>c.Content.Length) / 4 < maxSingleRequestTokenCount &&
+				input.Content.Length / 4 + chatMessages.Sum(c => c.Content.Length) / 4 < maxSingleRequestTokenCount &&
 				arg.ParseResult.GetValueForOption(AiAssistorCommands.StrategyOption) !=
 				FileRequestStrategy.SingleRequestForFile;
 
 			if (!addToExisting && hasContent && !allTokensSpend)
 			{
 				(tokenCount, costCent) =
-					await AskAi(chatMessages, model, openAiService, response, tokenCount, costCent);
+					await AskAi(chatMessages, model, openAiService, codeReviewComments, tokenCount, costCent);
 				if (maxTotalRequestTokenCount > 0 && tokenCount >= maxTotalRequestTokenCount)
 				{
 					// We already expended all tokens, we can't ask anymore.
@@ -137,7 +137,7 @@ internal class AiCommenter
 		// The last message that may be partially constructed.
 		if (!allTokensSpend)
 		{
-			(tokenCount, costCent) = await AskAi(chatMessages, model, openAiService, response, tokenCount,
+			(tokenCount, costCent) = await AskAi(chatMessages, model, openAiService, codeReviewComments, tokenCount,
 				costCent);
 		}
 
@@ -146,50 +146,83 @@ internal class AiCommenter
 			Console.WriteLine("All tokens spend - some parts of the PR were skipped.");
 		}
 
-		await CreateAiCommentInDevOps(model, tokenCount, costCent, response, gitClient, projectName, repositoryName,
+		await CreateAiCommentInDevOps(model, tokenCount, costCent, codeReviewComments, gitClient, projectName,
+			repositoryName,
 			pullRequestId, cancellationToken);
 	}
 
 	private static async Task CreateAiCommentInDevOps(Models.Model model, int tokenCount, double costCent,
-		StringBuilder response,
+		List<AiCodeReviewComment> codeReviewComments,
 		GitHttpClient gitClient, string projectName, string repositoryName, int pullRequestId,
 		CancellationToken cancellationToken)
 	{
-		StringBuilder commentBuilder = new();
-		commentBuilder.AppendLine($"## AI Suggestions ({model} - {tokenCount} token {costCent:F1} cent)");
-		commentBuilder.AppendLine(response.ToString());
-
-		Comment comment = new()
-		{
-			CommentType = CommentType.Text,
-			Content = commentBuilder.ToString()
-		};
-
 		List<GitPullRequestCommentThread>? threads = await gitClient.GetThreadsAsync(projectName,
 			repositoryId: repositoryName, pullRequestId: pullRequestId, cancellationToken: cancellationToken);
 
 		GitPullRequestCommentThread? aiCommentThread =
 			threads.FirstOrDefault(t => t.Properties?.Any(p => p.Key == "ChatGtpConversation") == true);
-		if (aiCommentThread != null)
+		if (aiCommentThread == null)
 		{
-			comment.ParentCommentId = aiCommentThread.Comments.First().Id;
-			await gitClient.CreateCommentAsync(comment, projectName, repositoryId: repositoryName,
-				pullRequestId: pullRequestId, threadId: aiCommentThread.Id, cancellationToken: cancellationToken);
-		}
-		else
-		{
-			GitPullRequestCommentThread commentThread = new GitPullRequestCommentThread
+			aiCommentThread = new GitPullRequestCommentThread
 			{
-				Comments = new List<Comment>
-				{
-					comment
-				},
+				Comments =
+				[
+					new Comment()
+					{
+						CommentType = CommentType.Text,
+						Content =
+							$"## AI Suggestions ({model} - {tokenCount} token {costCent:F1} cent)\r\n" +
+							"The following comments are generated by an AI model. Please like those that were useful."
+					}
+				],
+
 				Status = CommentThreadStatus.Active,
 				Properties = new PropertiesCollection { { "ChatGtpConversation", "true" } }
 			};
 
-			await gitClient.CreateThreadAsync(commentThread, projectName, repositoryId: repositoryName,
+			aiCommentThread = await gitClient.CreateThreadAsync(aiCommentThread, projectName,
+				repositoryId: repositoryName,
 				pullRequestId: pullRequestId, cancellationToken: cancellationToken);
+		}
+
+		foreach (AiCodeReviewComment codeReviewComment in codeReviewComments)
+		{
+			Comment comment = new Comment
+			{
+				CommentType = CommentType.Text,
+				Content = ToContentString(codeReviewComment)
+			};
+
+			if (comment.Content != null)
+			{
+				comment.ParentCommentId = aiCommentThread.Comments.First().Id;
+				await gitClient.CreateCommentAsync(comment, projectName, repositoryId: repositoryName,
+					pullRequestId: pullRequestId, threadId: aiCommentThread.Id, cancellationToken: cancellationToken);
+			}
+		}
+	}
+
+	private static string? ToContentString(AiCodeReviewComment codeReviewComment)
+	{
+		if (!codeReviewComment.Suggestions.Any())
+		{
+			return null;
+		}
+
+		if (codeReviewComment.Filename == string.Empty)
+		{
+			return codeReviewComment.Suggestions[0];
+		}
+		else
+		{
+			StringBuilder sb = new();
+			sb.AppendLine($"## {codeReviewComment.Filename}");
+			foreach (string suggestion in codeReviewComment.Suggestions)
+			{
+				sb.AppendLine($"- {suggestion}");
+			}
+
+			return sb.ToString();
 		}
 	}
 
@@ -261,13 +294,15 @@ internal class AiCommenter
 	}
 
 	private static async Task<(int tokenCount, double costCent)> AskAi(List<ChatMessage> chatMessagesStart,
-		Models.Model model, OpenAIService openAiService, StringBuilder response, int tokenCount, double costCent)
+		Models.Model model, OpenAIService openAiService, List<AiCodeReviewComment> comments, int tokenCount,
+		double costCent)
 	{
-		List<ChatMessage> chatMessages = new(chatMessagesStart);
+		List<ChatMessage> chatMessages = [..chatMessagesStart];
 
 		double tokenPromptCostInCent = GetTokenCosts(model, out double tokenCompletionCostInCent);
 
-		chatMessages.Add(ChatMessage.FromUser("Now please provide a summary of the changes. Ignore unchanged code, it is only provided for context, do not provide a review yet."));
+		chatMessages.Add(ChatMessage.FromUser(
+			"Now please provide a summary of the changes. Ignore unchanged code, it is only provided for context, do not provide a review yet."));
 
 		int count = 0;
 		bool success = false;
@@ -294,13 +329,24 @@ internal class AiCommenter
 						"""
 						Now provide constructive feedback on the changes.
 						Ignore code that was not changed in this PR.
-						Give between one and three suggestions per file with the most relevant first.
-						Use markdown as the output if possible. Output should be formatted like:
-						## [fullPathFilename]
-							- [suggestion 1 text]
-							- [suggestion 2 text]
-						...
-						Focus on the changed parts only, don't mention line numbers and don't be too nit-picky.
+						Give a minimum of one and a maximum of five suggestions per file with the most relevant first.
+						Output as a JSON in the format
+						{
+							"reviews": [
+								{
+									"filename": "fullPathFilename",
+									"suggestions": [
+										"suggestion 1 text",
+										"suggestion 2 text"
+									]
+								},
+								...
+						    ]
+						}
+
+						Do not add additional text beyond the JSON.
+
+						Focus on the changed parts, don't mention line numbers and don't be too nit-picky.
 						"""));
 
 					completionResult = await openAiService.ChatCompletion.CreateCompletion(
@@ -314,8 +360,7 @@ internal class AiCommenter
 					{
 						Console.WriteLine(completionResult.Choices.First().Message.Content);
 						chatMessages.Add(completionResult.Choices.First().Message);
-						response.AppendLine();
-						response.AppendLine(completionResult.Choices.First().Message.Content);
+						comments.AddRange(AiCodeReviewCommentParser.Parse(completionResult.Choices.First().Message.Content));
 						tokenCount += completionResult.Usage.TotalTokens;
 						costCent += tokenPromptCostInCent * completionResult.Usage.PromptTokens / 1000.0;
 						costCent += tokenCompletionCostInCent * (completionResult.Usage.CompletionTokens ?? 0) / 1000.0;
@@ -432,7 +477,7 @@ internal class AiCommenter
 						case ChangeType.Deleted:
 							diffBuilder.AppendLine("-" + line.Text);
 							break;
-						default: 
+						default:
 							diffBuilder.AppendLine(" " + line.Text);
 							break;
 					}
